+++ src/test/ExSln2/TranspileRunner/Program.cs
@@ -1,6 +1,7 @@
 using finlang;
 using finlang.Transpiler;
 using hal;
+using issue58;
 using System.Text.RegularExpressions;
 
 string thisDir = PathHelpers.GetThisDir();
@@ -10,7 +11,7 @@ string slnName = "ExSln2.sln";
 string outDir = slnDir + "/c99/gen";
 string projectName = "LedBlinker";
 
-Environment.SetEnvironmentVariable(CTranspiler.ENV_VAR_TRANSPILER_DEBUG_TYPE, nameof(FuncPtrEx2));
+Environment.SetEnvironmentVariable(CTranspiler.ENV_VAR_TRANSPILER_DEBUG_TYPE, nameof(VtableReturnsObjPointerEx));
 
 Console.WriteLine("Transpiling " + projectName + " fin/C# project...");